+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/DriveConstants.java
@@ -65,7 +65,7 @@ public class DriveConstants {
      * small and gradually increase them later after everything is working. All distance units are
      * inches.
      */
-    public static double MAX_VEL = 35;
+    public static double MAX_VEL = 45;
     public static double MAX_ACCEL = 30;
     public static double MAX_ANG_VEL = 7;
     public static double MAX_ANG_ACCEL = 3;
